+++ tests/integration_test.rs
@@ -71,6 +71,24 @@ fn test_stream() {
     assert_eq!(response, "HTTP/1.1 200 Ok\r\nContent-Type: text/event-stream\r\n\r\n: initial data\nHello.\nIs there anybody in there?\nJust nod if you can hear me.\n");
 }
 
+#[test]
+fn test_request_with_path_and_query_params() {
+    let server = TestServer::new().unwrap();
+    let resource = server.create_resource("/user/{userId}?filter=*&version=1");
+
+    resource
+        .status(Status::OK)
+        .header("Content-Type", "application/json")
+        .body(r#"{"id": 123, "userId": "{path.userId}", "filter": "{query.filter}", "v": {query.version}}"#);
+
+    let response = request(server.port(), "/user/superUser?filter=all&version=1", "GET");
+
+    assert_eq!(
+        response,
+        "HTTP/1.1 200 Ok\r\nContent-Type: application/json\r\n\r\n{\"id\": 123, \"userId\": \"superUser\", \"filter\": \"all\", \"v\": 1}"
+    );
+}
+
 fn request(port: u16, uri: &str, method: &str) -> String {
     let stream = open_stream(port, uri, method);