+++ CHANGELOG.md
@@ -2,7 +2,20 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
-## DRAFT
+## Unreleased
+
+The only breaking change in this release is that `resource.body("")` now can reference path and query parameters. What it means is that if `{path.<param_name>}` or `{query.<param_name>}` is provided, they will be translated to the values received in the request.
+
+### Added
+
+- Support to query and path parameters.
+- Support to using query and path parameters in body response.
+- Support to dynamic body.
+
+### Changed
+
+- Renamed some internal variables and methods in Resource.
+- Changed how server matches resources.
 
 ### Fixed